cilium: auto-discovery pod subnets for ENI IPAM
When encryption is enabled with an IPAM using a sinlge IP pool for the
entire cluster instead of per node we need to know those CIDRs. This is
required so we can place a decrypt/encrypt rules that covers the subnet.

In this patch lookup the CIDRs from the routeInfo and use those.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>